requirements.txt
black==20.8b1
coverage==5.3
Django==3.1.2
django-cors-headers==3.5.0
django-environ==0.4.5
django-extensions==3.0.9
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
drf-yasg==1.17.1
Pillow==8.0.0
